Compensating for the DaqMX/NiSync bug where DaqMx Export Signal will steal Trig0 from NiSync Create Clock.  Added a configuration where the DaqMX esported signal can be explicitly routed if needed:
Exported clock in the digitizer has been disabled so we can deal with this issue in a future commit.

A different issue in Sync Module.  When the reference clock looses lock, SyncLoopHandler was disabling the timed triggers.  This was not needed and caused an error when lock returned. Timed triggering is no longer disabled when entering unlocked state.